==> applied-insights/case-studies/posts/2023/08/21/_code/coleridge_supporting_code/baseline_model/00_read_data_from_DB.ipynb <==
## Read Data from Database

This notebook connects to the ADRF database and downloads all the data to raw_data folder and shows:
 - Table locations and how to load the data from the SQL database
 - How to save tables to a CSV file
 
This is only for demonstration purpose showing how to connect database and use SQL query to read data. In practice, one could directly use or preprocess data from the query instead of saving and reading the data. Also, we will involve the core column in this baseline model pipeline. For detailed documentation, please refer to the IRI/FNDDS/PPC data dictionary for columns specifications.

import sqlalchemy
import pandas as pd

host = '<HOST>'
DB_iri = 'ds_usda_iri'
DB_fndds = 'ds_usda_fndds'

# Connect to FNDDS database
connection = "mssql+pyodbc://{}/{}?driver=ODBC+Driver+17+for+SQL+Server?;Trusted_Connection=Yes".format(host, DB_fndds)
conn = sqlalchemy.create_engine(connection)

### Part 1: Reading FNDDS Data
We will extract EC code and EC description from these tables.

# Create main food description table
query_main = '''
SELECT * FROM dbo.mainfooddesc1516
'''
mainfooddesc1516 = pd.read_sql(query_main, conn)
mainfooddesc1516.to_csv('raw_data/mainfooddesc1516.csv', index=False)

mainfooddesc1516.head()

# Create additional food description table
query_add = '''
SELECT * FROM dbo.addfooddesc1516
'''
addfooddesc1516 = pd.read_sql(query_add, conn)
addfooddesc1516.to_csv('raw_data/addfooddesc1516.csv', index=False)

addfooddesc1516.head()

# Create ingredient description table
query_ingred = '''
SELECT * FROM dbo.fnddsingred1516
'''
fnddsingred1516 = pd.read_sql(query_ingred, conn)
fnddsingred1516.to_csv('raw_data/fnddsingred1516.csv', index=False)

fnddsingred1516.head()

### Part 2: Reading IRI Data
We will extract UPC code and UPC description from this table。

# Connect to IRI database
connection = "mssql+pyodbc://{}/{}?driver=ODBC+Driver+17+for+SQL+Server?;Trusted_Connection=Yes".format(host, DB_iri)
conn = sqlalchemy.create_engine(connection)

# Create iri table
# The IRI data is relatively big, this cell will take several miuntes to run
# In order to access pd_master_all table, the table name should be changed to vw_pd_master_2016
query_iri = '''
SELECT * FROM dbo.vw_pd_pos_2016
WHERE year = 2015 or year = 2016
'''
pd_pos_all = pd.read_sql(query_iri, conn)
pd_pos_all.to_csv('raw_data/pd_pos_all1516.csv', index=False)

pd_pos_all.head()

### Part 3：Reading PPC Data
This is the crosswalk table between UPC code and EC code. It serves as the ground truth for the competition.

# Create ppc table 
query_ppc = '''
SELECT * FROM dbo.ppc20152016
'''
ppc20152016 = pd.read_sql(query_ppc, conn)
ppc20152016.to_csv('raw_data/ppc20152016.csv', index=False)

ppc20152016.head()
